Add files via upload
This is my code, I have done my best to make it as I just a beginner in this GATE. I am trying to make code based on description on journal "Simulation study of a DOI-based PET-Compton imaging system for positron emitters". And I currently stuck here. I plan to make sensitivity as the output but the explanation on GATE collaboration is to complicated for me. Please help me...